fix(🎨): fix race condition in SkiaDomView (#2823)
fixes #2814

The race condition would occur when prop update in the view would trigger a redraw request, then the DOM  would also emit a  redraw request itself would get canceled due to the pending redraw. That first redraw request however, is not using the updated DOM version.

We are rearchitecturing this part of the system at the moment, where such issues won't be possible anymore. However this should be good fix for now.